refactor: Extract shared gh command utilities to reduce code duplication
Fixes SonarCloud code duplication issue (6.7% -> under 3%).

Changes:
- Created scripts/lib/gh-command.js with shared security utilities
- Extracted executeGhCommand, validateIssueNumber, validatePRNumber, validateTag
- Updated check-orphaned-issues.js to use shared utilities
- Maintains all security fixes from DMCP-SEC-002

Benefits:
- Eliminates code duplication between scripts
- Centralizes security patterns in one location
- Makes future script development easier and more secure
- Passes SonarCloud quality gate

Related: DMCP-SEC-002 security fixes
